Add default value for encrypted password
This isn't necessary–it was added for practice with migrations.

Devise sets null: false, default: "" because:

In the event that you're updating an existing table, the DB engine will try to populate that new column with NULL in each row. In such cases, you must override this behavior with default: "", or else it will conflict with null: false and the migraiton will fail.

https://stackoverflow.com/questions/44904357/why-would-you-set-null-false-default-on-a-required-db-column